chore: rename params to sshnp_params
