+++ packages/noports_core/lib/config_repository.dart
@@ -1,5 +1,5 @@
 library sshnp_core_config_repository;
 
-export 'src/sshnp/params/config_file_repository.dart';
-export 'src/sshnp/params/config_key_repository.dart';
-export 'src/sshnp/params/sshnp_params.dart';
+export 'src/sshnp/sshnp_params/config_file_repository.dart';
+export 'src/sshnp/sshnp_params/config_key_repository.dart';
+export 'src/sshnp/sshnp_params/sshnp_params.dart';
