+++ packages/noports_core/lib/src/sshnp/sshnp.dart
@@ -6,7 +6,7 @@ import 'package:at_utils/at_utils.dart';
 import 'package:dartssh2/dartssh2.dart';
 import 'package:meta/meta.dart';
 import 'package:noports_core/src/common/default_args.dart';
-import 'package:noports_core/src/sshnp/params/sshnp_params.dart';
+import 'package:noports_core/src/sshnp/sshnp_params/sshnp_params.dart';
 import 'package:noports_core/src/common/supported_ssh_clients.dart';
 import 'package:noports_core/src/common/utils.dart';
 import 'package:noports_core/src/sshnp/sshnp_impl.dart';
