+++ packages/noports_core/lib/src/sshnp/sshnp_params/config_file_repository.dart
@@ -1,8 +1,8 @@
 import 'dart:io';
 
 import 'package:noports_core/src/common/utils.dart';
-import 'package:noports_core/src/sshnp/params/sshnp_params.dart';
-import 'package:noports_core/src/sshnp/params/sshnp_arg.dart';
+import 'package:noports_core/src/sshnp/sshnp_params/sshnp_params.dart';
+import 'package:noports_core/src/sshnp/sshnp_params/sshnp_arg.dart';
 import 'package:path/path.dart' as path;
 
 class ConfigFileRepository {
