+++ packages/noports_core/lib/sshnp.dart
@@ -2,4 +2,4 @@ library noports_core_sshnp;
 
 export 'src/sshnp/sshnp.dart';
 export 'src/sshnp/sshnp_result.dart';
-export 'src/sshnp/params/sshnp_params.dart';
+export 'src/sshnp/sshnp_params/sshnp_params.dart';
